move sidebar into component
