+++ frontend/src/components/SideBar.css
@@ -1,0 +1,59 @@
+/* Sidebar.css */
+
+.sidebar {
+    height: 100%;
+    width: 250px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    background-color: #7D5F50;
+    padding-top: 20px;
+  }
+  
+.sidebar ul {
+    list-style-type: none;
+    padding: 0;
+}
+
+.sidebar-control {
+    width: 30px;
+    height: 30px;
+    margin-bottom: 10px;
+    display: flex;
+    margin-left: auto;
+    margin-right: 20px;
+    margin-top: -20px;
+    cursor: pointer;
+}
+
+.sidebar-logo {
+    width: 150px;
+    height: 150px;
+    margin-bottom: 10px;
+    /* add img to center */
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+
+}
+
+.sidebar-icon {
+    width: 25px;
+    height: 25px;
+    margin-right: 10px;
+}
+
+.sidebar li {
+    padding: 8px;
+    text-align: left;
+    margin-left: auto;
+    color: white;
+    cursor: pointer;
+    display: flex;
+}
+
+.sidebar li:hover {
+    background-color: #D1D1D1;
+    color: black;
+}
+
